manta: Dexopt is disabled by default
So remove the flags for it

Pre-optimization is controlled by the build option WITH_DEXPREOPT.
Before the L release, this was enabled by default in “user” builds.
Starting in L, this option is opt-in and needs to be enabled
in the product configuration such as a device’s BoardConfig.mk file.

Change-Id: If62a256aafde68bed3c99d7173d882f5f8fc8df3

Conflicts:
	BoardConfig.mk